Move repeat spec to `src/spec/repeat.ts`
