+++ src/compile/repeat.ts
@@ -1,7 +1,7 @@
 import {Config} from '../config';
 import * as log from '../log';
-import {Repeat} from '../repeat';
 import {NormalizedRepeatSpec} from '../spec';
+import {Repeat} from '../spec/repeat';
 import {VgLayout} from '../vega.schema';
 import {BaseConcatModel} from './baseconcat';
 import {buildModel} from './buildmodel';
