--- src/repeat.ts
@@ -1,11 +1,0 @@
-export interface Repeat {
-  /**
-   * Vertical repeated views.
-   */
-  row?: string[];
-
-  /**
-   * Horizontal repeated views.
-   */
-  column?: string[];
-}
